requirements.txt
duckduckgo2==0.242
wolframalpha==1.0.2
requests==2.2.1
wikipedia==1.1
svn+http://pygoogle.googlecode.com/svn/trunk#egg=pygoogle
urlnorm==1.1.2
lxml==3.3.3
Flask-API==0.6.2
Flask==0.10.1
langid==1.1.4dev
gevent==1.0
pytest==2.5.2
ipdb==0.8
ipython==2.0.0
-e git+https://github.com/gimbel0893/bingsearch.git@v0.4#egg=bingsearch